Line Following Robot
This project is made for Embedded System course's project which is used Arduino to code simple robots that contains at least 5 specialibility. This car typed robot follows the line with reading the color black and white, it is sensitive to colors,  so when it seem black their wheels turn on and start trip in black line when it reads white , car stops moving.